Size

+++ 0x00-hello_world/6-size.c
@@ -9,9 +9,9 @@ int intType;
 float floatType;
 double doubleType;
 char charType;
-printf("Size of an int: %lu byte(s)\n", sizeof(intType));
-printf("Size of a float: %lu byte(s)\n", sizeof(floatType));
-printf("Size of a double: %lu byte(s)\n", sizeof(doubleType));
-printf("Size of a char: %lu byte(s)\n", sizeof(charType));
+printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(intType));
+printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(floatType));
+printf("Size of a double: %lu byte(s)\n", (unsigned long)sizeof(doubleType));
+printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(charType));
 return (0);
 }
